+++ test/test_helper.exs
@@ -1,1 +1,1 @@
-ExUnit.start()
+ExUnit.start(exclude: [:live_gpu, :live_cpu_heavy, :live_apple, :network])